Keep canvas in place when hiding or showing docks
When hitting tab by default. At least mostly. There can still be a very
slight shift if zoomed at a subpixel scale, but the view no longer jerks
around anyway. I think this used to work at one point, but then broke?
Not sure when or how though.

+++ src/desktop/mainwindow.h
@@ -126,7 +126,7 @@ public:
 signals:
 	void hostSessionEnabled(bool enabled);
 	void windowReplacementFailed(MainWindow *win);
-	void viewShifted(int deltaX, int deltaY);
+	void viewShifted(qreal deltaX, qreal deltaY);
 
 public slots:
 	// Triggerable actions
